perf: centralize thumbnail logic and use background thumbnails on travel highlights page
